Create Netflix Data Analysis with Python
Among the biggest suppliers of internet streaming services is Netflix. Due to its massive subscriber base, it gathers a tremendous amount of data. I'm going to expose you to a data science project using Python to analyze data from Netflix in this article. Because Netflix has shifted its business model from on-demand DVD movie rental to focusing heavily on the production of their original series, the platform has continually focused on changing business needs, making it possible for us to examine a large amount of data and models from Netflix.

Some of the most important tasks that we can analyze from Netflix data are:
understand what content is available, understand the similarities between the content, understand the network between actors and directors, what exactly Netflix is focusing on, and sentiment analysis of content available on Netflix.